# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      confirm_password:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
    required:
    - confirm_password
    - password
    type: object
...
  dto.CreateUserRequest:
    properties:
      email:
        example: budi@gudang.com
        type: string
      password:
        example: rahasia123
        minLength: 6
        type: string
      role:
        enum:
        - admin
        - planner
        - operator
        example: planner
        type: string
      username:
        example: budi
        maxLength: 50
        minLength: 4
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
      role:
        enum:
        - admin
        - planner
        - operator
        type: string
      username:
        maxLength: 50
        minLength: 4
        type: string
    type: object
  dto.UserProfileResponse:
    properties:
      address:
        type: string
      avatar_url:
        type: string
      date_of_birth:
        type: string
      full_name:
        type: string
      gender:
        type: string
      phone_number:
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      profile:
        $ref: '#/definitions/dto.UserProfileResponse'
      role:
        type: string
      username:
        type: string
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/response.ErrorDetail'
        type: array
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      source: {}
    type: object
  response.Meta:
    properties:
      count:
        type: integer
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for Load Stuffing Calculator application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Load Stuffing Calculator API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of users.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user with the specified role. Requires admin privileges.
      parameters:
      - description: User Creation Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user by ID. Requires admin privileges.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieves user details by their ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates an existing user. Requires admin privileges.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      description: Changes the password for a specific user. Requires admin privileges.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password Change Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Change user password
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"